+++ src/Oro/Bundle/PaymentBundle/Action/PaymentTransactionCaptureAction.php
@@ -21,7 +21,7 @@ class PaymentTransactionCaptureAction extends AbstractPaymentMethodAction
         parent::configureOptionsResolver($resolver);
 
         $resolver
-            ->remove(['object', 'amount', 'currency', 'paymentMethod'])
+            ->remove(['object', 'amount', 'currency'])
             ->setRequired([self::OPTION_PAYMENT_TRANSACTION])
             ->addAllowedTypes(
                 self::OPTION_PAYMENT_TRANSACTION,
@@ -35,7 +35,7 @@ class PaymentTransactionCaptureAction extends AbstractPaymentMethodAction
         parent::configureValuesResolver($resolver);
 
         $resolver
-            ->remove(['object', 'amount', 'currency', 'paymentMethod'])
+            ->remove(['object', 'amount', 'currency'])
             ->setRequired([self::OPTION_PAYMENT_TRANSACTION])
             ->addAllowedTypes(self::OPTION_PAYMENT_TRANSACTION, PaymentTransaction::class);
     }