+++ src/Oro/Bundle/PaymentBundle/Tests/Unit/Action/PaymentTransactionCaptureActionTest.php
@@ -185,7 +185,6 @@ class PaymentTransactionCaptureActionTest extends \PHPUnit\Framework\TestCase
             [['object' => 'someValue']],
             [['amount' => 'someAmount']],
             [['currency' => 'someCurrency']],
-            [['paymentMethod' => 'somePaymentMethod']],
         ];
     }