+++ src/Oro/Bundle/PaymentBundle/Tests/Unit/Action/PaymentTransactionRefundActionTest.php
@@ -181,7 +181,6 @@ class PaymentTransactionRefundActionTest extends \PHPUnit\Framework\TestCase
             [['object' => 'someValue']],
             [['amount' => 'someAmount']],
             [['currency' => 'someCurrency']],
-            [['paymentMethod' => 'somePaymentMethod']],
         ];
     }